Fix: 피드 fetch시 전체 타입일 때만 캐시 삭제하도록 수정

+++ app/src/main/java/com/android04/godfisherman/data/repository/FeedRepository.kt
@@ -19,7 +19,6 @@ import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.launch
 import java.util.*
 import javax.inject.Inject
-import javax.inject.Singleton
 
 class FeedRepository @Inject constructor(
     private val localDataSource: FeedDataSource.LocalDataSource,
@@ -28,8 +27,10 @@ class FeedRepository @Inject constructor(
     @ApplicationScope private val externalScope: CoroutineScope
 ) {
     private fun fetchPagingData(type: Type): Flow<PagingData<FeedData>> {
-        externalScope.launch {
-            localDataSource.deleteAll()
+        if (type == Type.ALL) {
+            externalScope.launch {
+                localDataSource.deleteAll()
+            }
         }
 
         return Pager(PagingConfig(pageSize = 2)) { FeedRemotePagingSource(type) }.flow
